Refactor and fix pyflakes errors in sympy.assumptions.
